+++ react/lib/components/stencil-generated/index.ts
@@ -1,0 +1,68 @@
+/* eslint-disable */
+/* tslint:disable */
+/* auto-generated react proxies */
+import { createReactComponent } from './react-component-lib';
+
+import type { JSX } from 'wcs-core';
+
+
+
+export const WcsAccordion = /*@__PURE__*/createReactComponent<JSX.WcsAccordion, HTMLWcsAccordionElement>('wcs-accordion');
+export const WcsAccordionContent = /*@__PURE__*/createReactComponent<JSX.WcsAccordionContent, HTMLWcsAccordionContentElement>('wcs-accordion-content');
+export const WcsAccordionHeader = /*@__PURE__*/createReactComponent<JSX.WcsAccordionHeader, HTMLWcsAccordionHeaderElement>('wcs-accordion-header');
+export const WcsAccordionPanel = /*@__PURE__*/createReactComponent<JSX.WcsAccordionPanel, HTMLWcsAccordionPanelElement>('wcs-accordion-panel');
+export const WcsActionBar = /*@__PURE__*/createReactComponent<JSX.WcsActionBar, HTMLWcsActionBarElement>('wcs-action-bar');
+export const WcsApp = /*@__PURE__*/createReactComponent<JSX.WcsApp, HTMLWcsAppElement>('wcs-app');
+export const WcsBadge = /*@__PURE__*/createReactComponent<JSX.WcsBadge, HTMLWcsBadgeElement>('wcs-badge');
+export const WcsButton = /*@__PURE__*/createReactComponent<JSX.WcsButton, HTMLWcsButtonElement>('wcs-button');
+export const WcsCard = /*@__PURE__*/createReactComponent<JSX.WcsCard, HTMLWcsCardElement>('wcs-card');
+export const WcsCardBody = /*@__PURE__*/createReactComponent<JSX.WcsCardBody, HTMLWcsCardBodyElement>('wcs-card-body');
+export const WcsCheckbox = /*@__PURE__*/createReactComponent<JSX.WcsCheckbox, HTMLWcsCheckboxElement>('wcs-checkbox');
+export const WcsComNav = /*@__PURE__*/createReactComponent<JSX.WcsComNav, HTMLWcsComNavElement>('wcs-com-nav');
+export const WcsComNavCategory = /*@__PURE__*/createReactComponent<JSX.WcsComNavCategory, HTMLWcsComNavCategoryElement>('wcs-com-nav-category');
+export const WcsComNavSubmenu = /*@__PURE__*/createReactComponent<JSX.WcsComNavSubmenu, HTMLWcsComNavSubmenuElement>('wcs-com-nav-submenu');
+export const WcsCounter = /*@__PURE__*/createReactComponent<JSX.WcsCounter, HTMLWcsCounterElement>('wcs-counter');
+export const WcsDivider = /*@__PURE__*/createReactComponent<JSX.WcsDivider, HTMLWcsDividerElement>('wcs-divider');
+export const WcsDropdown = /*@__PURE__*/createReactComponent<JSX.WcsDropdown, HTMLWcsDropdownElement>('wcs-dropdown');
+export const WcsDropdownDivider = /*@__PURE__*/createReactComponent<JSX.WcsDropdownDivider, HTMLWcsDropdownDividerElement>('wcs-dropdown-divider');
+export const WcsDropdownHeader = /*@__PURE__*/createReactComponent<JSX.WcsDropdownHeader, HTMLWcsDropdownHeaderElement>('wcs-dropdown-header');
+export const WcsDropdownItem = /*@__PURE__*/createReactComponent<JSX.WcsDropdownItem, HTMLWcsDropdownItemElement>('wcs-dropdown-item');
+export const WcsEditableField = /*@__PURE__*/createReactComponent<JSX.WcsEditableField, HTMLWcsEditableFieldElement>('wcs-editable-field');
+export const WcsError = /*@__PURE__*/createReactComponent<JSX.WcsError, HTMLWcsErrorElement>('wcs-error');
+export const WcsField = /*@__PURE__*/createReactComponent<JSX.WcsField, HTMLWcsFieldElement>('wcs-field');
+export const WcsFieldContent = /*@__PURE__*/createReactComponent<JSX.WcsFieldContent, HTMLWcsFieldContentElement>('wcs-field-content');
+export const WcsFieldLabel = /*@__PURE__*/createReactComponent<JSX.WcsFieldLabel, HTMLWcsFieldLabelElement>('wcs-field-label');
+export const WcsFooter = /*@__PURE__*/createReactComponent<JSX.WcsFooter, HTMLWcsFooterElement>('wcs-footer');
+export const WcsFormField = /*@__PURE__*/createReactComponent<JSX.WcsFormField, HTMLWcsFormFieldElement>('wcs-form-field');
+export const WcsGalactic = /*@__PURE__*/createReactComponent<JSX.WcsGalactic, HTMLWcsGalacticElement>('wcs-galactic');
+export const WcsGalacticMenu = /*@__PURE__*/createReactComponent<JSX.WcsGalacticMenu, HTMLWcsGalacticMenuElement>('wcs-galactic-menu');
+export const WcsGrid = /*@__PURE__*/createReactComponent<JSX.WcsGrid, HTMLWcsGridElement>('wcs-grid');
+export const WcsGridColumn = /*@__PURE__*/createReactComponent<JSX.WcsGridColumn, HTMLWcsGridColumnElement>('wcs-grid-column');
+export const WcsGridCustomCell = /*@__PURE__*/createReactComponent<JSX.WcsGridCustomCell, HTMLWcsGridCustomCellElement>('wcs-grid-custom-cell');
+export const WcsGridPagination = /*@__PURE__*/createReactComponent<JSX.WcsGridPagination, HTMLWcsGridPaginationElement>('wcs-grid-pagination');
+export const WcsHeader = /*@__PURE__*/createReactComponent<JSX.WcsHeader, HTMLWcsHeaderElement>('wcs-header');
+export const WcsHint = /*@__PURE__*/createReactComponent<JSX.WcsHint, HTMLWcsHintElement>('wcs-hint');
+export const WcsHorizontalStepper = /*@__PURE__*/createReactComponent<JSX.WcsHorizontalStepper, HTMLWcsHorizontalStepperElement>('wcs-horizontal-stepper');
+export const WcsIcon = /*@__PURE__*/createReactComponent<JSX.WcsIcon, HTMLWcsIconElement>('wcs-icon');
+export const WcsInput = /*@__PURE__*/createReactComponent<JSX.WcsInput, HTMLWcsInputElement>('wcs-input');
+export const WcsLabel = /*@__PURE__*/createReactComponent<JSX.WcsLabel, HTMLWcsLabelElement>('wcs-label');
+export const WcsListItem = /*@__PURE__*/createReactComponent<JSX.WcsListItem, HTMLWcsListItemElement>('wcs-list-item');
+export const WcsListItemProperties = /*@__PURE__*/createReactComponent<JSX.WcsListItemProperties, HTMLWcsListItemPropertiesElement>('wcs-list-item-properties');
+export const WcsListItemProperty = /*@__PURE__*/createReactComponent<JSX.WcsListItemProperty, HTMLWcsListItemPropertyElement>('wcs-list-item-property');
+export const WcsMatIcon = /*@__PURE__*/createReactComponent<JSX.WcsMatIcon, HTMLWcsMatIconElement>('wcs-mat-icon');
+export const WcsModal = /*@__PURE__*/createReactComponent<JSX.WcsModal, HTMLWcsModalElement>('wcs-modal');
+export const WcsNativeSelect = /*@__PURE__*/createReactComponent<JSX.WcsNativeSelect, HTMLWcsNativeSelectElement>('wcs-native-select');
+export const WcsNav = /*@__PURE__*/createReactComponent<JSX.WcsNav, HTMLWcsNavElement>('wcs-nav');
+export const WcsNavItem = /*@__PURE__*/createReactComponent<JSX.WcsNavItem, HTMLWcsNavItemElement>('wcs-nav-item');
+export const WcsProgressBar = /*@__PURE__*/createReactComponent<JSX.WcsProgressBar, HTMLWcsProgressBarElement>('wcs-progress-bar');
+export const WcsProgressRadial = /*@__PURE__*/createReactComponent<JSX.WcsProgressRadial, HTMLWcsProgressRadialElement>('wcs-progress-radial');
+export const WcsRadio = /*@__PURE__*/createReactComponent<JSX.WcsRadio, HTMLWcsRadioElement>('wcs-radio');
+export const WcsRadioGroup = /*@__PURE__*/createReactComponent<JSX.WcsRadioGroup, HTMLWcsRadioGroupElement>('wcs-radio-group');
+export const WcsSelect = /*@__PURE__*/createReactComponent<JSX.WcsSelect, HTMLWcsSelectElement>('wcs-select');
+export const WcsSelectOption = /*@__PURE__*/createReactComponent<JSX.WcsSelectOption, HTMLWcsSelectOptionElement>('wcs-select-option');
+export const WcsSpinner = /*@__PURE__*/createReactComponent<JSX.WcsSpinner, HTMLWcsSpinnerElement>('wcs-spinner');
+export const WcsSwitch = /*@__PURE__*/createReactComponent<JSX.WcsSwitch, HTMLWcsSwitchElement>('wcs-switch');
+export const WcsTab = /*@__PURE__*/createReactComponent<JSX.WcsTab, HTMLWcsTabElement>('wcs-tab');
+export const WcsTabs = /*@__PURE__*/createReactComponent<JSX.WcsTabs, HTMLWcsTabsElement>('wcs-tabs');
+export const WcsTextarea = /*@__PURE__*/createReactComponent<JSX.WcsTextarea, HTMLWcsTextareaElement>('wcs-textarea');
+export const WcsTooltip = /*@__PURE__*/createReactComponent<JSX.WcsTooltip, HTMLWcsTooltipElement>('wcs-tooltip');